+++ src/main/java/tfar/classicbar/impl/overlays/vanilla/HarmType.java
@@ -1,0 +1,5 @@
+package tfar.classicbar.impl.overlays.vanilla;
+
+public enum HarmType {
+    NONE, POISON, WITHER, FROZEN
+}